[fix] 디바이스의 크기에 맞춰 Dialog의 크기를 지정하는 함수를 BaseDialogFragment로 이동함
