+++ app/src/main/java/com/bodan/maplecalendar/presentation/BaseDialogFragment.kt
@@ -1,12 +1,15 @@
 package com.bodan.maplecalendar.presentation
 
+import android.content.Context
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
+import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.Window
+import android.view.WindowManager
 import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
@@ -41,5 +44,30 @@ abstract class BaseDialogFragment<T : ViewDataBinding>(private val layoutId: Int
 
     protected fun showToastMessage(message: String) {
         Toast.makeText(requireContext(), message, Toast.LENGTH_LONG).show()
+    }
+
+    protected fun Context.dialogFragmentResize(
+        dialogFragment: DialogFragment,
+        width: Float,
+        height: Float
+    ) {
+        val windowManager = getSystemService(Context.WINDOW_SERVICE) as WindowManager
+        val window = dialogFragment.dialog?.window
+        if (Build.VERSION.SDK_INT < 30) {
+            val displayWidth = resources.displayMetrics.widthPixels
+            val displayHeight = resources.displayMetrics.heightPixels
+
+            val x = (displayWidth * width).toInt()
+            val y = (displayHeight * height).toInt()
+
+            window?.setLayout(x, y)
+        } else {
+            val rect = windowManager.currentWindowMetrics.bounds
+
+            val x = (rect.width() * width).toInt()
+            val y = (rect.height() * height).toInt()
+
+            window?.setLayout(x, y)
+        }
     }
 }
