+++ app/src/main/java/com/bodan/maplecalendar/presentation/lobby/SearchDateFragment.kt
@@ -1,11 +1,7 @@
 package com.bodan.maplecalendar.presentation.lobby
 
-import android.content.Context
-import android.os.Build
 import android.os.Bundle
 import android.view.View
-import android.view.WindowManager
-import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
 import com.bodan.maplecalendar.R
@@ -50,31 +46,6 @@ class SearchDateFragment :
         with(binding.vpSearchDate) {
             adapter = customCalendarAdapter
             setCurrentItem(CustomCalendarAdapter.START_POSITION, false)
-        }
-    }
-
-    private fun Context.dialogFragmentResize(
-        dialogFragment: DialogFragment,
-        width: Float,
-        height: Float
-    ) {
-        val windowManager = getSystemService(Context.WINDOW_SERVICE) as WindowManager
-        val window = dialogFragment.dialog?.window
-        if (Build.VERSION.SDK_INT < 30) {
-            val displayWidth = resources.displayMetrics.widthPixels
-            val displayHeight = resources.displayMetrics.heightPixels
-
-            val x = (displayWidth * width).toInt()
-            val y = (displayHeight * height).toInt()
-
-            window?.setLayout(x, y)
-        } else {
-            val rect = windowManager.currentWindowMetrics.bounds
-
-            val x = (rect.width() * width).toInt()
-            val y = (rect.height() * height).toInt()
-
-            window?.setLayout(x, y)
         }
     }
 }
